requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
astroid==2.9.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
black==19.10b0
bleach==4.0.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography==36.0.0
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.1.2
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
docutils==0.17.1
entrypoints==0.3
Flask==2.0.2
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.45.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.2
intervaltree==3.1.0
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.2
jsonschema==3.2.0
jupyterlab-pygments==0.1.2
kaleido==0.2.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numexpr==2.7.3
numpy
numpydoc==1.1.0
oauthlib==3.2.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.4
platformdirs==2.4.0
plotly==5.4.0
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.12.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyorbital==1.7.1
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
pytz==2021.3
pyu2f==0.1.5
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==2.8.1
QtAwesome==1.0.3
QtPy==1.10.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.1
retrying==1.3.3
rope==0.21.1
rsa==4.8
Rtree==0.9.7
scipy==1.7.3
Send2Trash==1.8.0
setuptools==58.0.4
sip==5.0.0
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.27
tenacity==8.0.1
testpath==0.5.0
textdistance==4.2.1
three-merge==0.1.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
ujson==4.0.2
uritemplate==4.1.1
urllib3==1.26.7
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.3
yapf==0.31.0
yarl==1.7.2
zipp==3.6.0
